+++ skidl/pyspice.py
@@ -5,6 +5,7 @@
 """
 Import this file to reconfigure SKiDL for doing SPICE simulations.
 """
+
 from __future__ import (  # isort:skip
     absolute_import,
     division,
@@ -14,9 +15,16 @@ from __future__ import (  # isort:skip
 
 from future import standard_library
 
-from skidl import *
-
-from .tools import SKIDL, SPICE
+from .skidl import set_default_tool, generate_netlist, reset, lib_search_paths, generate_svg
+from .net import Net
+from .schlib import SchLib
+from .part import TEMPLATE, Part
+from .group import subcircuit
+from .tools.spice import node, Parameters, XspiceModel
+from .package import package
+from .bus import Bus
+from .tools import SKIDL, SPICE, KICAD
+from .logger import active_logger
 
 standard_library.install_aliases()
 
@@ -25,10 +33,12 @@ try:
     from PySpice import *
     from PySpice.Unit import *
 
-    from .libs.pyspice_sklib import *
 except ImportError:
-    pass
+    active_logger.raise_(NotImplementedError, "PySpice does not support Python 2, so SPICE simulation is not possible.")
+
 else:
+    from .libs.pyspice_sklib import *
+
     _splib = SchLib("pyspice", tool=SKIDL)  # Read-in the SPICE part library.
 
     set_default_tool(SPICE)  # Set the library format for reading SKiDL libraries.
